pyproject.toml
[tool.poetry]
name = "ecoindex-monorepo"
version = "0.1.0"
description = ""
authors = ["Vincent Vatelot <vincent.vatelot@ik.me>"]
readme = "README.md"
packages = [
    { include = "development" },
    { include = "ecoindex/backend", from = "bases" },
    { include = "ecoindex/cli", from = "bases" },
    { include = "ecoindex/compute", from = "components" },
    { include = "ecoindex/data", from = "components" },
    { include = "ecoindex/models", from = "components" },
    { include = "ecoindex/scraper", from = "components" },
    { include = "ecoindex/scripts", from = "components" },
    { include = "ecoindex/utils", from = "components" },
    { include = "ecoindex/worker", from = "bases" },
    { include = "ecoindex/database", from = "components" },
    { include = "ecoindex/exceptions", from = "components" },
    { include = "ecoindex/config", from = "components" },
]

[tool.poetry.dependencies]
aiofile = "^3.8.8"
loguru = "^0.7.2"
pydantic = "^2.4.2"
python = ">=3.10,<3.13"
requests = "^2.31.0"
pydantic-settings = "^2.0.3"

[tool.poetry.group.scraper.dependencies]
pillow = "^10.1.0"
playwright = "^1.39.0"
playwright-stealth = "^1.0.6"
python-slugify = "^8.0.1"

[tool.poetry.group.cli.dependencies]
click-spinner = "^0.1.10"
jinja2 = "^3.1.2"
matplotlib = "^3.8.0"
pandas = "^2.1.2"
pyyaml = "^6.0.1"
rich = "^13.6.0"
scrapy = "^2.11.0"
typer = "^0.9.0"

[tool.poetry.group.api.dependencies]
aiosqlite = "^0.19.0"
alembic = "^1.12.1"
celery = "^5.3.4"
redis = "^5.0.1"
sqlmodel = { git = "https://github.com/AntonDeMeester/sqlmodel.git" }

[tool.poetry.group.api-backend.dependencies]
fastapi = "^0.104.0"
fastapi-health = "^0.4.0"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.1"
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
update-values = "ecoindex.scripts:update_values"
ecoindex-cli = "ecoindex.cli.app:app"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
